+++ migrations/19_add_compressed_flag_down.sql
@@ -1,0 +1,2 @@
+ALTER TABLE media DROP COLUMN compressed;
+ALTER TABLE thumbnails DROP COLUMN compressed;