+++ types/media.go
@@ -14,6 +14,7 @@ type Media struct {
 	Location    string
 	CreationTs  int64
 	Quarantined bool
+	Compressed  bool
 }
 
 type MinimalMedia struct {