+++ types/object_info.go
@@ -1,7 +1,5 @@
 package types
 
 type ObjectInfo struct {
-	Location   string
-	Sha256Hash string
-	SizeBytes  int64
+	Location string
 }