+++ types/thumbnail.go
@@ -17,6 +17,7 @@ type Thumbnail struct {
 	Location    string
 	CreationTs  int64
 	Sha256Hash  string
+	Compressed  bool
 }
 
 type StreamedThumbnail struct {